inventory_management/requirements.txt
asgiref
Django
djangorestframework
sqlparse
psycopg2-binary
typing_extensions
django-leaflet
django-import-export
pytest
pytest-django
coverage